third_party/blink/web_tests/fast/forms/color/color-picker-tab-navigation-expected.txt

Popup opened.
EYE-DROPPER
COLOR-SELECTION-RING
RVALUECONTAINER
GVALUECONTAINER
BVALUECONTAINER
FORMAT-TOGGLER
COLOR-SELECTION-RING
FORMAT-TOGGLER clicked. Active color format changed.
COLOR-SELECTION-RING
EYE-DROPPER
COLOR-SELECTION-RING
HVALUECONTAINER
SVALUECONTAINER
LVALUECONTAINER
FORMAT-TOGGLER
FORMAT-TOGGLER clicked. Active color format changed.
COLOR-SELECTION-RING
EYE-DROPPER
COLOR-SELECTION-RING
HEXVALUECONTAINER
FORMAT-TOGGLER
TEST COMPLETE
